Define `MuchDecimal` mixin
This makes the `MuchDecimal` a mixin that can be used to define
attributes that are decimals stored as integers. This allows easily
setting up attributes that are decimals but store their values as
integers. The decimals can be used for presentation/display while
the integer can be used for math operations. This is particularly
useful when working with currency.

This brings in much-plugin and makes the main module `MuchDecimal`
a plugin. The plugin simply adds a class method
`decimal_as_integer` that allows setting up decimal as integer
attributes. The class method takes an attribute name and a set of
options. The source attribute and a precision are the possible
options. The source attributes specifies the accessor that the
integer will be written to. The precision is the number of decimal
places or how precise you want the integer to be. Both of these are
defaulted if they aren't provided.

The class method defines a reader and writer. The readers reads
from the source accessor expecting an integer and converting it to
its decimal using the precision. The writer expects a decimal and
converts it to an integer using the precision and writes it to the
source accessor. Both of these handle `nil` and non numeric values
appropriately.

Finally, this pulls the conversion logic, decimal to integer and
integer to decimal, into singleton methods on the main module.
This allows their logic to be used directly. This is particularly
useful for testing that an attribute is using much-decimal.