Add `TestHelpers`
This adds test helpers that can be used to test that a class is
using `MuchDecimal` for an attribute. This is to make the gem
friendlier to use and so users don't have to dig into the internals
of much-decimal just to test something that uses it.

The `TestHelpers` provide an `assert_decimal_as_integer` helper
method that expects a subject, attribute and options. The subject
should be an object that has mixed in `MuchDecimal` and defined
a decimal as integer attribute. The test helper will write a
decimal to the attribute and ensure the source and attribute
return the correct value.